Organize commands in to folders
Break out individual commands from src/commands.ts in to individual
files that export their main command function.
