+++ src/commands/dependencies/useLocal.ts
@@ -1,0 +1,65 @@
+//===----------------------------------------------------------------------===//
+//
+// This source file is part of the VS Code Swift open source project
+//
+// Copyright (c) 2021-2024 the VS Code Swift project authors
+// Licensed under Apache License v2.0
+//
+// See LICENSE.txt for license information
+// See CONTRIBUTORS.txt for the list of VS Code Swift project authors
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+//===----------------------------------------------------------------------===//
+
+import * as vscode from "vscode";
+import { FolderEvent, WorkspaceContext } from "../../WorkspaceContext";
+import { createSwiftTask } from "../../tasks/SwiftTaskProvider";
+import { executeTaskWithUI } from "../utilities";
+
+/**
+ * Use local version of package dependency
+ *
+ * equivalent of `swift package edit --path <localpath> identifier
+ * @param identifier Identifier for dependency
+ * @param ctx workspace context
+ */
+export async function useLocalDependency(identifier: string, ctx: WorkspaceContext) {
+    const currentFolder = ctx.currentFolder;
+    if (!currentFolder) {
+        return;
+    }
+    const folders = await vscode.window.showOpenDialog({
+        canSelectFiles: false,
+        canSelectFolders: true,
+        canSelectMany: false,
+        defaultUri: currentFolder.folder,
+        openLabel: "Select",
+        title: "Select folder",
+    });
+
+    if (!folders) {
+        return;
+    }
+    const folder = folders[0];
+    const task = createSwiftTask(
+        ["package", "edit", "--path", folder.fsPath, identifier],
+        "Edit Package Dependency",
+        {
+            scope: currentFolder.workspaceFolder,
+            cwd: currentFolder.folder,
+            prefix: currentFolder.name,
+        },
+        ctx.toolchain
+    );
+
+    const success = await executeTaskWithUI(
+        task,
+        `Use local version of ${identifier}`,
+        currentFolder,
+        true
+    );
+    if (success) {
+        ctx.fireEvent(currentFolder, FolderEvent.resolvedUpdated);
+    }
+}
